refactor(#2050 W1.3②): move spawn_one from api/mod.rs to agent_ops
`spawn_one` (the SPAWN / CREATE_TEAM choke point) lived in the api server file
`api/mod.rs` but is an agent-spawn primitive — it already calls
`agent_ops::remove_metadata` and is invoked from `api/handlers/{instance,team}`
via `crate::api::spawn_one`. Move it (verbatim, behavior-unchanged) to its
cohesive home `agent_ops`, made `pub`, and repoint the three callers to
`crate::agent_ops::spawn_one`. Dropped the now-unused `self` from api/mod.rs's
`use crate::agent::{...}`.

§10.5: the TUI-serve thread spawn rode `api/mod.rs`'s legacy
`EXEMPTED_LEGACY_FILES` grandfather; agent_ops is in-scope for the
spawn-rationale audit, so the move upgrades it with a real
`// fire-and-forget:` rationale (the exemption list is meant to shrink, not
grow) — net improvement, not a new exemption.

The api/mod.rs `spawn_one_resolves_preset_submit_key` test stays put: it
exercises `Backend::from_command`/`preset`, not the moved symbol.


Agend-Agent: fixup-dev
Agend-Branch: refactor/extract-spawn-one-to-agent-ops
Agend-Issued-At: 2026-06-12T08:04:08.006051+00:00
